refact(ngdoc): refactor inline tags to be able to return a function to run later
We were running inline tags after all tags and plugins had completed. This is necessary because
some in-line tags need meta-data that is extracted from other tags in order to work; but this meant
that we had to search the processed doc properties, recursively, for all inline tags, even in the
original text.

This refactoring allows in-line tag handler to return either a string or a function.  If the handler
returns a string it will replace the tag in the text. If the handler returns a function the function
will be stored and then run at the end of the ngdoc processing.  The result of running this function
will then replace the in-line tag in the text, wherever it appears in the doc.