fix(examples processor): only parse the content once
The processor was walking all the properties but this was causing many
duplicate examples to be parsed.  Now it only looks at the content property
and this is done before any further parsing of the doc is done.

Importantly this means that you can't add properties on the fly later on
containing <example> tags and expect them to be extracted.