fix(examples-generate processor): remove hard-coded script link
The link to the angular.js files was hard-coded to work with the example
here but failed when used in the real angular.js project since the path to
angular.js was different.  This change makes that path configurable and is
a better solution all round. You can now provide a list of common files
in your package config that will be used in every example.

For example, you could do this:

```
config.set('processing.examples.commonFiles', {
    scripts: [ '../../angular.js' ],
    stylesheets: []
  });
```

Now in every example you will get a link to the angular.js file two folders
above the example's index.html