chore(version): bump to 0.1.0
Hopefully, this will help our CI server with getting the latest version.
Versions with 0 for their major version number have special semantics,
which mean that 0.0.y -> 0.0.y+1 is deemed a breaking change so npm won't
automatically update the version.

By moving to 0.1.x we should be better off as 0.1.x -> 0.1.x+1 is not
deemed a breaking change.